requirements.txt
python==3.9
fastapi
uvicorn[standard]  
python-multipart 
python-jose[cryptography]
passlib[bcrypt] 
sqlalchemy
fastapi-mail